Created abstract class
