+++ contact/src/main/java/com/example/contact/model/Organization.java
@@ -5,89 +5,30 @@ import java.time.LocalDateTime;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+//import javax.persistence.Entity;
+//import javax.persistence.GeneratedValue;
+//import javax.persistence.GenerationType;
+//import javax.persistence.Id;
 import javax.persistence.Table;
 
+
 @Entity
-
 @Table(name = "Organization")
-public class Organization {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "Org_ID")
-	private Long id;
-
-	@Column(name = "Organization_Name")
-	private String name;
-
-	@Column(name = "Organization_number")
-	private String phoneNumber;
-
-	@Column(name = "Organization_email")
-	private String email;
-
-	@Column(name = "created_at")
-	private LocalDateTime createdAt;
+public class Organization extends Contact{
 
 	@Column(name = "Website")
 	private String website;
 
-	public Organization() {
-		super();
-		this.createdAt = LocalDateTime.now();
-	}
-
-	public Organization(Long id, String name, String phoneNumber, String email, 
+	public Organization(Long id, String name, String phoneNumber, String email, LocalDateTime createdAt,
 			String website) {
-		super();
-		this.id = id;
-		this.name = name;
-		this.phoneNumber = phoneNumber;
-		this.email = email;
+		super(id, name, phoneNumber, email, createdAt);
 		this.website = website;
 	}
 
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public LocalDateTime getCreatedAt() {
-		return createdAt;
-	}
-
-	public void setCreatedAt(LocalDateTime createdAt) {
-		this.createdAt = createdAt;
+	
+	public Organization() {
+		super();
+	
 	}
 
 	public String getWebsite() {
@@ -97,11 +38,7 @@ public class Organization {
 	public void setWebsite(String website) {
 		this.website = website;
 	}
+	
 
-	@Override
-	public String toString() {
-		return "Organization [id=" + id + ", name=" + name + ", phoneNumber=" + phoneNumber + ", email=" + email
-				+ ", createdAt=" + createdAt + ", website=" + website + "]";
-	}
 
 }
