+++ contact/src/main/java/com/example/contact/model/Person.java
@@ -2,96 +2,23 @@ package com.example.contact.model;
 
 import java.time.LocalDateTime;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Entity
-
 @Table(name = "Person")	
-public class Person {
-	
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "person_id")
-	private Long id;
-	
-	@Column(name = "full_name")
-	private String name;
-	
-	@Column(name = "phone_number")
-	private String phoneNumber;
-	
-	@Column(name = "email")
-	private String email;
-	
+public class Person extends Contact {
 
-	@Column(name = "created_at")
-	private LocalDateTime createdAt;
-
+	public Person(Long id, String name, String phoneNumber, String email, LocalDateTime createdAt) {
+		super(id, name, phoneNumber, email, createdAt);
+		
+	}
 
 	public Person() {
 		super();
-		this.createdAt = LocalDateTime.now();
 	
 	}
 	
-	public Person(Long id, String name, String phoneNumber, String email) {
-		super();
-		this.id = id;
-		this.name = name;
-		this.phoneNumber = phoneNumber;
-		this.email = email;
-		
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public LocalDateTime getCreatedAt() {
-		return createdAt;
-	}
-
-	public void setCreatedAt(LocalDateTime createdAt) {
-		this.createdAt = createdAt;
-	}
-
-	@Override
-	public String toString() {
-		return "Person [id=" + id + ", name=" + name + ", phoneNumber=" + phoneNumber + ", email=" + email
-				+ ", createdAt=" + createdAt + "]";
+	
 	}
 		
-}
